Integrated HERE SDK 4.12.10 and Flutter 3.3.2
Removed deprecations for HERE SDK 4.12.10, deprecated MapUpdater
functions are still used (checkMapUpdate and performMapUpdate).

+++ lib/route_preferences/dropdown_widget.dart
@@ -55,7 +55,7 @@ class DropdownWidget extends StatelessWidget {
           .toList(),
       onChanged: (value) {
         if (value != null) {
-          onChanged(value as int);
+          onChanged(value);
         }
       },
     );
